feat: migrate verifyInitData util to TypeScript
