+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Файл `release.yml` собирает образ и выкладывает его на Railway.
 
 - Включён строгий режим TypeScript и добавлен план миграции из JavaScript (`docs/typescript_migration_plan.md`).
-- Переписаны на TypeScript утилиты `userLink`, `formatTask`, `validate` и `haversine`.
+- Переписаны на TypeScript утилиты `userLink`, `formatTask`, `validate`, `haversine` и `verifyInitData`.
 
 - Исправлено подключение middleware `logging` и `metrics` в API.
 - Убраны лишние аргументы маршрута `/api/v1/users`, исключающего ошибку `argument handler must be a function`.
