--- bot/src/utils/verifyInitData.js
@@ -1,23 +1,0 @@
-// Проверяет подпись initData из Telegram WebApp
-// Модули: crypto, config
-const crypto = require('crypto');
-const { botToken } = require('../config');
-
-module.exports = function verifyInitData(initData) {
-  const params = new URLSearchParams(initData);
-  const hash = params.get('hash');
-  params.delete('hash');
-  const dataCheckString = [...params.entries()]
-    .sort(([a], [b]) => a.localeCompare(b))
-    .map(([k, v]) => `${k}=${v}`)
-    .join('\n');
-  const secretKey = crypto
-    .createHmac('sha256', 'WebAppData')
-    .update(botToken)
-    .digest();
-  const hmac = crypto
-    .createHmac('sha256', secretKey)
-    .update(dataCheckString)
-    .digest('hex');
-  return hmac === hash;
-};
