+++ docs/typescript_migration_plan.md
@@ -11,11 +11,11 @@
 | `bot/tests` | 46 |
 | `scripts` и корневые конфиги | 13 |
 
-Переведены на TypeScript: `userLink`, `formatTask`, `validate`, `haversine`.
+Переведены на TypeScript: `userLink`, `formatTask`, `validate`, `haversine`, `verifyInitData`.
 
 ### Высокий приоритет — сервер `bot/src`
 
-- `utils`: verifyInitData.js, setTokenCookie.js, accessMask.js, rateLimiter.js, formatUser.js
+- `utils`: setTokenCookie.js, accessMask.js, rateLimiter.js, formatUser.js
 - `bot`: bot.js, messages.js
 - `controllers`: maps.js, optimizer.js, routes.js
 - `middleware`: taskAccess.js, checkRole.js
